requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.8
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
gunicorn==21.2.0
alembic==1.13.2
PyMySQL==1.1.0
psycopg2-binary==2.9.9
